Microservices/transactionsMicro/requirements.txt
django==3.1.12
djangorestframework==3.11.0
djongo==1.3.7
pika==1.3.2